added new function to calculate sweeps between e2 measurements and infinite loop for the algorithm

+++ source/gauge_fixing.h
@@ -1,7 +1,7 @@
 #ifndef GAUGEFIXING_H
 #define GAUGEFIXING_H
 
-unsigned short SU3_gauge_fix(double complex *U, const unsigned short config);
+unsigned SU3_gauge_fix(double complex *U, const unsigned short config);
 
 // void SU3_print_gaugefixed_U(double complex * U, double complex *U_aux, char filename[max_length_name]);
